+++ src/shared/BoardAssessors.ts
